Multiple changes to refactor the error/panic handling.
When sfnt.Parse(..) is called it only checks the general structure is correct, not that each table is valid. As it goes it stores the offsets within the font so parsing of each table can be done later. Later when indivisual tables are used, they are parsed (once) and a error can occur.