Move font to cmd/font; merge commands into it.
This change simplifies the overall project and makes it more consistent.

Removing the commands package reduces the public facing API surface,
making it so that whatever API is left is higher quality, more useful.
This should be helpful to people who are discovering the project for the
first time and reading its documentation.

Resolves #10.